Extract create_item from mutations/files.rs
